+++ android/app/src/main/AndroidManifest.xml
@@ -25,9 +25,9 @@
 
     <application
         android:allowBackup="true"
-        android:icon="@mipmap/云体测_icon"
+        android:icon="@mipmap/icon"
         android:label="@string/app_name"
-        android:roundIcon="@mipmap/云体测_icon_round"
+        android:roundIcon="@mipmap/icon_round"
         android:supportsRtl="true"
         android:taskAffinity=""
         android:theme="@style/AppTheme"
