+++ android/app/src/main/java/com/cloudsports/actiondetect/actiondetect/MainViewModel.kt
@@ -1,0 +1,65 @@
+/*
+ * Copyright 2023 The TensorFlow Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.cloudsports.actiondetect.actiondetect
+
+import androidx.lifecycle.ViewModel
+
+/**
+ *  This ViewModel is used to store pose landmarker helper settings
+ */
+class MainViewModel : ViewModel() {
+
+    private var _model = PoseLandmarkerHelper.MODEL_POSE_LANDMARKER_FULL
+    private var _delegate: Int = PoseLandmarkerHelper.DELEGATE_CPU
+    private var _minPoseDetectionConfidence: Float =
+        PoseLandmarkerHelper.DEFAULT_POSE_DETECTION_CONFIDENCE
+    private var _minPoseTrackingConfidence: Float =
+        PoseLandmarkerHelper.DEFAULT_POSE_TRACKING_CONFIDENCE
+    private var _minPosePresenceConfidence: Float =
+        PoseLandmarkerHelper.DEFAULT_POSE_PRESENCE_CONFIDENCE
+
+    val currentDelegate: Int get() = _delegate
+    val currentModel: Int get() = _model
+    val currentMinPoseDetectionConfidence: Float
+        get() =
+            _minPoseDetectionConfidence
+    val currentMinPoseTrackingConfidence: Float
+        get() =
+            _minPoseTrackingConfidence
+    val currentMinPosePresenceConfidence: Float
+        get() =
+            _minPosePresenceConfidence
+
+    fun setDelegate(delegate: Int) {
+        _delegate = delegate
+    }
+
+    fun setMinPoseDetectionConfidence(confidence: Float) {
+        _minPoseDetectionConfidence = confidence
+    }
+
+    fun setMinPoseTrackingConfidence(confidence: Float) {
+        _minPoseTrackingConfidence = confidence
+    }
+
+    fun setMinPosePresenceConfidence(confidence: Float) {
+        _minPosePresenceConfidence = confidence
+    }
+
+    fun setModel(model: Int) {
+        _model = model
+    }
+}
