+++ android/app/src/main/java/com/cloudsports/actiondetect/actiondetect/OverlayView.kt
@@ -1,0 +1,161 @@
+/*
+ * Copyright 2023 The TensorFlow Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.cloudsports.actiondetect.actiondetect
+
+import android.content.Context
+import android.graphics.Canvas
+import android.graphics.Color
+import android.graphics.Paint
+import android.os.Build
+import android.util.AttributeSet
+import android.view.View
+import androidx.annotation.RequiresApi
+import androidx.core.content.ContextCompat
+import com.cloudsports.actiondetect.actiondetect.algorithm.PullUpActionCount
+import com.cloudsports.actiondetect.actiondetect.debug.ToastDebug
+import com.google.mediapipe.tasks.vision.core.RunningMode
+import com.google.mediapipe.tasks.vision.poselandmarker.PoseLandmarker
+import com.google.mediapipe.tasks.vision.poselandmarker.PoseLandmarkerResult
+import org.json.JSONArray
+import kotlin.math.max
+import kotlin.math.min
+
+class OverlayView(context: Context?, attrs: AttributeSet?) :
+    View(context, attrs) {
+
+    private var results: PoseLandmarkerResult? = null
+    private var pointPaint = Paint()
+    private var linePaint = Paint()
+
+    private var scaleFactor: Float = 1f
+    private var imageWidth: Int = 1
+    private var imageHeight: Int = 1
+
+    public var lastCount = 0
+    public var count = 0
+    private val pullUp_action_count = PullUpActionCount(context)
+
+    private var toast: ToastDebug? = ToastDebug(context)
+    init {
+        initPaints()
+    }
+
+    fun clear() {
+        results = null
+        pointPaint.reset()
+        linePaint.reset()
+        invalidate()
+        initPaints()
+    }
+
+    private fun initPaints() {
+        linePaint.color =
+            ContextCompat.getColor(context!!, R.color.mp_color_primary)
+        linePaint.strokeWidth = LANDMARK_STROKE_WIDTH
+        linePaint.style = Paint.Style.STROKE
+
+        pointPaint.color = Color.YELLOW
+        pointPaint.strokeWidth = LANDMARK_STROKE_WIDTH
+        pointPaint.style = Paint.Style.FILL
+    }
+
+    override fun draw(canvas: Canvas) {
+        super.draw(canvas)
+
+        // 将count绘制到屏幕上
+        val paint = Paint()
+        paint.color = Color.RED
+        paint.textSize = 100f
+        canvas.drawText(count.toString(), 100f, 100f, paint)
+
+        results?.let { poseLandmarkerResult ->
+            for(landmark in poseLandmarkerResult.landmarks()) {
+                val landmarkArray = JSONArray()
+                for(normalizedLandmark in landmark) {
+                    canvas.drawPoint(
+                        normalizedLandmark.x() * imageWidth * scaleFactor,
+                        normalizedLandmark.y() * imageHeight * scaleFactor,
+                        pointPaint
+                    )
+                }
+                PoseLandmarker.POSE_LANDMARKS.forEach {
+                    canvas.drawLine(
+                        poseLandmarkerResult.landmarks().get(0).get(it!!.start())
+                            .x() * imageWidth * scaleFactor,
+                        poseLandmarkerResult.landmarks().get(0).get(it.start())
+                            .y() * imageHeight * scaleFactor,
+                        poseLandmarkerResult.landmarks().get(0).get(it.end())
+                            .x() * imageWidth * scaleFactor,
+                        poseLandmarkerResult.landmarks().get(0).get(it.end())
+                            .y() * imageHeight * scaleFactor,
+                        linePaint
+                    )
+                }
+            }
+
+        }
+    }
+
+    @RequiresApi(Build.VERSION_CODES.O)
+    fun setResults(
+        poseLandmarkerResults: PoseLandmarkerResult,
+        imageHeight: Int,
+        imageWidth: Int,
+        runningMode: RunningMode = RunningMode.IMAGE
+    ) {
+        results = poseLandmarkerResults
+
+        this.imageHeight = imageHeight
+        this.imageWidth = imageWidth
+
+        // 将results.landmarks()转化为Array<DoubleArray>
+        val poseLandmarks = Array<DoubleArray>(33) { DoubleArray(3) }
+
+        if (poseLandmarkerResults.landmarks().isNotEmpty()) {
+            for (i in 0..32) {
+                poseLandmarks[i][0] = poseLandmarkerResults.landmarks()[0][i].x().toDouble()
+                poseLandmarks[i][1] = poseLandmarkerResults.landmarks()[0][i].y().toDouble()
+                poseLandmarks[i][2] = poseLandmarkerResults.landmarks()[0][i].z().toDouble()
+            }
+            lastCount = count
+            count = pullUp_action_count(poseLandmarks, imageHeight, imageWidth)
+            if (lastCount != count) {
+                toast?.show("count: $count")
+            }
+        }
+
+
+
+
+        scaleFactor = when (runningMode) {
+            RunningMode.IMAGE,
+            RunningMode.VIDEO -> {
+                min(width * 1f / imageWidth, height * 1f / imageHeight)
+            }
+            RunningMode.LIVE_STREAM -> {
+                // PreviewView is in FILL_START mode. So we need to scale up the
+                // landmarks to match with the size that the captured images will be
+                // displayed.
+                max(width * 1f / imageWidth, height * 1f / imageHeight)
+            }
+        }
+        invalidate()
+    }
+
+    companion object {
+        private const val LANDMARK_STROKE_WIDTH = 12F
+    }
+}
