+++ android/app/src/main/java/com/cloudsports/actiondetect/actiondetect/algorithm/PoseSample.kt
@@ -1,0 +1,7 @@
+package com.cloudsports.actiondetect.actiondetect.algorithm
+
+class PoseSample(name: String, val landmarks: Array<DoubleArray>, val className: String,
+                 val embedding: Array<DoubleArray>
+) {
+    private val _name = name
+}
