+++ android/app/src/main/java/com/cloudsports/actiondetect/actiondetect/algorithm/PoseSampleOutlier.kt
@@ -1,0 +1,7 @@
+package com.cloudsports.actiondetect.actiondetect.algorithm
+
+class PoseSampleOutlier(sample: PoseSample, detected_class: Map<String, Int>, all_classes: Map<String, Int>) {
+    private val _sample = sample
+    private val _detectedClass = detected_class
+    private val _allClasses = all_classes
+}
