+++ android/app/src/main/res/mipmap-hdpi/icon.png
