+++ android/app/src/main/res/mipmap-mdpi/icon.png
