+++ android/app/src/main/res/mipmap-xxhdpi/icon.png
